Adding first layouts of the engine and the board
Choosing an array of integers as the data representation and
initialising useful indexes (Such as North and South starting point +
store).

Then filling the board with the given number of seeds and simple allow
to print the result as a String.

The Engine responsability will be to hold the board, the game logic
(Such as player turns, capture conditions, ...)