Implement capture of opponent's seeds
If the last seed fall in an empty (owned) houses and the opponent's house is not
empty, "capture" both houses' seeds in the store.

At the moment, the capture condition logic is in the board but really
should be in the engine. I am thinking about implementing the "Empty
capture" variant of Kalah and that should be the responsability of the
Engine.

Also added methods to get a player's score.