Fix Toggle Turn, refactor capture and add tests
Fair amount of refactoring once again but most importantly, I have
discovered that if the last seed fall in the player's store, he is
entitlted to a free turn!

As explained before, the capture condition logic has been moved to the
engine with the idea of maybe implementating the "Empty capture" later
on.