Add coverage check
Now there is a good range of tests, it's time to actually make sure
we're good. I think Jacoco will generate a Report in the target
directory.

What I like to add early is a check on the coverage that will fail the
build if too many, untested lines are added at once.