Implementing front end API
Ok. That was a big chunk of work and should have been split in more than
one commit to be safe.

Basically prepared all the API work that would allow a full play.
Standard REST API here, with Json responses, correct verbs and sensible
status code.

The Engine and Board God exception from before has been deprecated in
favour of more expressive ones.

Finally, I decided against using an actual database to store the Games
as it sounded overkill to bring Hibernate for a simple example.
However, I tried to use the smae Repository pattern as there would be
in case of Spring Data, just populating an in memory HasMap.